requirements.txt
# Telegram API
python-telegram-bot==20.3

# Парсинг RSS
feedparser==6.0.10

# OpenAI API
openai==1.2.3
aiohttp==3.8.5
requests==2.31.0

# Google Sheets API (только для утреннего поста)
gspread==5.7.2
google-auth==2.23.3

# Для корректной работы с датами/часами (если понадобится)
pytz==2023.3

# Для обработки HTML и экранирования
html5lib==1.1
telegram
pillow
